Переделываем логи

+++ bot/handlers/video.py
@@ -1,6 +1,7 @@
 import asyncio
 import uuid
 import os
+import traceback
 from pathlib import Path
 from aiogram import Router, Bot
 from aiogram.filters import Command
@@ -28,30 +29,44 @@ async def download_video(url: str) -> dict:
         return {"status": "error", "message": stderr.decode()}
     
 @video_router.message(Command("video"))
-async def cmd_video(message: Message, bot: Bot, url=None):
-    split_text = message.text.split()
-    processing_msg = await message.answer("⏳ Скачиваю, ждите")
+async def cmd_video(message: Message, bot: Bot):
+    command = "video"
+    url = None
+    file_path = None
+    processing_msg = None
 
-    if url is None and len(split_text) > 1 and split_text[1]:
-        url = split_text[1]
+    try:
+        if db.is_user_mediabanned(message.from_user.id):
+            await message.reply("❌ Вы заблокированы, это действие вам запрещено")
+            return
 
-    if url is not None:
-        result = await download_video(url)
-        if result["status"] == "success":
-            file = result["file_path"]
-            vid = FSInputFile(file)
-            await message.reply_video(vid, caption="📹 Вот ваше видео:")
+        processing_msg = await message.answer("⏳ Скачиваю, ждите")
 
-            process = await asyncio.create_subprocess_exec('rm', '-f', file)
-            await process.wait()
+        split_text = message.text.split()
+        if len(split_text) > 1 and split_text[1]:
+            url = split_text[1]
         else:
-            report_id = uuid.uuid4()
-            await message.reply(f"❌ Не удалось загрузить видео\nReport id: {report_id}")
-            await bot.send_message(os.getenv("OWNER_ID"), f"Report id: {report_id}\n\nMessage: {message.text}\n\nLogs: {result["message"]}")
-    else:
-        await message.reply("❌ Непредвиденная ошибка.")
+            return await message.reply("❌ Укажите URL видео в команде")
 
-    await processing_msg.delete()
+        result = await download_video(url)
+        if result["status"] != "success":
+            await error_report(message, bot, command, result["message"])
+            return
+
+        file_path = result["file_path"]
+        vid = FSInputFile(file_path)
+        await message.reply_video(vid, caption="📹 Вот ваше видео:")
+
+    except Exception:
+        error_traceback = traceback.format_exc()
+        await error_report(message, bot, command, error_traceback)
+        
+    finally:
+        if file_path and Path(file_path).exists():
+            process = await asyncio.create_subprocess_exec('rm', '-f', file_path)
+            await process.wait()
+        if processing_msg:
+            await processing_msg.delete()
 
 @video_router.message(Command("gif"))
 async def cmd_gif(message: Message, bot: Bot):
@@ -103,8 +118,9 @@ async def cmd_gif(message: Message, bot: Bot):
                 await message.reply_animation(gif)
         else:
             await message.reply("❌ Ошибка при конвертации.")
-    except Exception as e:
-        await error_report(message, bot, command, e)
+    except Exception:
+        er_traceback = traceback.format_exc()
+        await error_report(message, bot, command, er_traceback)
     finally:
         if input_path and input_path.exists():
             input_path.unlink(missing_ok=True)
